split faction components into their own individual files
